pyproject.toml
[tool.poetry]
name = "capstone-display"
version = "0.1.0"
description = ""
package-mode = false
authors = ["kmiya <miyak19@mytru.ca>"]

[tool.poetry.dependencies]
python = "^3.12"
msg-handler = { git = "https://github.com/TRU-Capstone-2026-WheelChock/wheelchoke-system-communication-system.git" }
pyyaml = "^6.0.3"
pyzmq = "^27.1.0"

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
pytest = ">=9.0.2,<10.0.0"
types-pyyaml = "^6.0.12.20250915"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["test"]
addopts = "-ra -q"
